Implemented find by period
