+++ docs/copyright.http
@@ -13,7 +13,7 @@ GET {{host}}/api/copyrights/id=1
 Accept: application/json
 
 > {%
-client.test("get copyright by id", function() {
+client.test("should find copyright by id", function() {
   client.assert(response.status === 200, "Response status is not 200");
 });
 %}
@@ -23,7 +23,7 @@ GET {{host}}/api/copyrights/company_id=1
 Accept: application/json
 
 > {%
-client.test("Request executed successfully", function() {
+client.test("should find all copyrights by company id", function() {
   client.assert(response.status === 200, "Response status is not 200");
 });
 %}
