+++ src/main/resources/application.properties
@@ -10,4 +10,4 @@ server.port=8080
 #
 logging.level.root=INFO
 logging.level.web=TRACE
-logging.level.com.example.offerdaysongs.service.CopyrightService=TRACE
+logging.level.com.example.offerdaysongs.service.CopyrightServiceImpl=TRACE
